+++ scripts/generate-openapi.ts
@@ -96,7 +96,7 @@ class OpenAPIGenerator {
     ['ldapFlatGeneric', 'Entities'],
     ['ldapBulkImport', 'Bulk Import'],
     ['james', 'Apache James Integration'],
-    ['calendarResources', 'Calendar Resources'],
+    ['calendar', 'Twake Calendar'],
     ['configApi', 'Configuration'],
     ['static', 'Static Files'],
     ['trash', 'Trash'],